Extract FactoryLogger.
